@@ -1,9 +1,9 @@
-package com.wk.guestpass.app.Adpaters;
+package com.wk.guestpass.app.adpaters;
 
 import android.content.Context;
 import android.graphics.Color;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.RecyclerView;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,7 +12,7 @@ import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.wk.guestpass.app.Models.ListModel;
+import com.wk.guestpass.app.models.ListModel;
 import com.wk.guestpass.app.R;
 import com.github.ivbaranov.mli.MaterialLetterIcon;
 
