@@ -1,17 +1,17 @@
-package com.wk.guestpass.app.Fragments;
+package com.wk.guestpass.app.fragments;
 
 import android.app.Dialog;
 import android.app.ProgressDialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SearchView;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.core.content.ContextCompat;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.SearchView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -31,18 +31,19 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.wk.guestpass.app.FragAdapters.UpcGuestAdapter;
-import com.wk.guestpass.app.Models.ListModel;
+import com.wk.guestpass.app.fragAdapters.TodaysGuestAdapter;
+import com.wk.guestpass.app.models.ListModel;
 import com.wk.guestpass.app.R;
-import com.wk.guestpass.app.Utilities.Config;
-import com.wk.guestpass.app.Utilities.RecyclerTouchListener;
-import com.wk.guestpass.app.Utilities.SessionManager;
+import com.wk.guestpass.app.utilities.Config;
+import com.wk.guestpass.app.utilities.RecyclerTouchListener;
+import com.wk.guestpass.app.utilities.SessionManager;
 import com.github.ybq.android.spinkit.style.CubeGrid;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-
+import com.crashlytics.android.Crashlytics;
+import io.fabric.sdk.android.Fabric;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -50,20 +51,19 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import com.crashlytics.android.Crashlytics;
-import io.fabric.sdk.android.Fabric;
+
 import es.dmoral.toasty.Toasty;
 import pl.droidsonroids.gif.GifImageView;
 
-public class UpcGuest extends Fragment {
+public class TodaysGuest extends Fragment {
 
-    private SessionManager session;
-    private String usersssid;
     private RecyclerView recyclerView;
-    private UpcGuestAdapter upcGuestAdapter;
+    private SessionManager session;
+    private String usersssid, userna;
+    private TextView dname, ddate, dsetime, dintime, dcntct, dguest, dpurpose;
+    private TodaysGuestAdapter todaysGuestAdapter;
     SwipeRefreshLayout mSwipeRefreshLayout;
     private ImageView nodata, logout, back, guestroles, expstmp;
-    private TextView dname, ddate, dsetime, dintime, dcntct, dguest, dpurpose;
     private List<ListModel> list = new ArrayList<>();
     private Dialog bottomSheetDialog;
     public static final String TAG = "MyTag";
@@ -79,20 +79,19 @@ public class UpcGuest extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         Fabric.with(getActivity(), new Crashlytics());
-        View v = inflater.inflate(R.layout.fragupcominguest, container, false);
-        session = new SessionManager(getActivity());
+        View v = inflater.inflate(R.layout.fragtodaysguest, container, false);
 
+        session = new SessionManager(getActivity());
         HashMap<String, String> users = session.getUserDetails();
         usersssid = users.get(SessionManager.KEY_ID);
-        nodata = (ImageView)v.findViewById(R.id.nodaata);
-        logout = v.findViewById(R.id.logout);
-        back  = v.findViewById(R.id.back);
 
+        nodata = (ImageView)v.findViewById(R.id.nodaata);
         progressBar = (ProgressBar)v. findViewById(R.id.progress);
         mainscreen=(RelativeLayout)v.findViewById(R.id.overmain);
         bgrnd=(RelativeLayout)v.findViewById(R.id.bgrnd);
 
-
+        logout = v.findViewById(R.id.logout);
+        back  = v.findViewById(R.id.back);
         searchView = (SearchView)v.findViewById(R.id.searchview);
         mSwipeRefreshLayout = (SwipeRefreshLayout) v.findViewById(R.id.swipeToRefresh);
         back.setOnClickListener(new View.OnClickListener() {
@@ -102,9 +101,9 @@ public class UpcGuest extends Fragment {
             }
         });
 
-        recyclerView = (RecyclerView)v.findViewById(R.id.UpcomingRecycler);
+        recyclerView = (RecyclerView)v.findViewById(R.id.TodaysRecycler);
         recyclerView.setHasFixedSize(false);
-     /*   RecyclerView.LayoutManager mLayoutManager = new LinearLayoutManager(getActivity());
+       /* RecyclerView.LayoutManager mLayoutManager = new LinearLayoutManager(getActivity());
         recyclerView.setLayoutManager(mLayoutManager);*/
         recyclerView.setLayoutManager(new GridLayoutManager(getActivity(), 2));
         recyclerView.setItemAnimator(new DefaultItemAnimator());
@@ -112,16 +111,18 @@ public class UpcGuest extends Fragment {
                 recyclerView, new RecyclerTouchListener.ClickListener() {
             @Override
             public void onClick(View view, int position) {
-                  bottomsheetdialog(position);
+                bottomsheetdialog(position);
             }
 
             @Override
             public void onLongClick(View view, final int position) {
                 final String stat=list.get(position).getGueststatus();
+                if(stat.equals("0")){
                     android.app.AlertDialog.Builder builder = new android.app.AlertDialog.Builder(getActivity());
                     builder.setMessage("Are you sure you want to delete");
                     builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
                         public void onClick(DialogInterface dialog, int id) {
+
                             deleteguest(list.get(position).getIds());
                             dialog.dismiss();
                         }
@@ -134,10 +135,21 @@ public class UpcGuest extends Fragment {
                     android.app.AlertDialog alertDialog = builder.create();
                     alertDialog.show();
                 }
-
+                else{
+                    Toasty.error(getActivity(), "Sorry!!! Data Can't be Deleted", Toast.LENGTH_SHORT, true).show();
+                }
+            }
         }));
-        upcGuestAdapter=new UpcGuestAdapter(getActivity(), list);
-        recyclerView.setAdapter(upcGuestAdapter);
+        todaysGuestAdapter=new TodaysGuestAdapter(getActivity(), list);
+        recyclerView.setAdapter(todaysGuestAdapter);
+
+        if(!usersssid.isEmpty()){
+            cubeGrid = new CubeGrid();
+            cubeGrid.setColor(getResources().getColor(R.color.colorPrimary));
+            cubeGrid.start();
+            progressBar.setIndeterminateDrawable(cubeGrid);
+            TodaysdataList();
+        }
 
         mSwipeRefreshLayout = (SwipeRefreshLayout)v.findViewById(R.id.swipeToRefresh);
         mSwipeRefreshLayout.setColorSchemeResources(R.color.colorAccent);
@@ -145,18 +157,10 @@ public class UpcGuest extends Fragment {
             @Override
             public void onRefresh() {
 
-                UpcomingList();
+                TodaysdataList();
                 mSwipeRefreshLayout.setRefreshing(false);
             }
         });
-
-        if(!usersssid.isEmpty()){
-            cubeGrid = new CubeGrid();
-            cubeGrid.setColor(getResources().getColor(R.color.colorPrimary));
-            cubeGrid.start();
-            progressBar.setIndeterminateDrawable(cubeGrid);
-            UpcomingList();
-        }
 
 
         logout.setOnClickListener(new View.OnClickListener() {
@@ -196,16 +200,17 @@ public class UpcGuest extends Fragment {
 
             @Override
             public boolean onQueryTextChange(String newText) {
-                if (upcGuestAdapter == null) {
+                if (todaysGuestAdapter == null) {
                     return false;
                 } else {
-                    upcGuestAdapter.getFilter().filter(newText);
+                    todaysGuestAdapter.getFilter().filter(newText);
                     return true;
                 }
             }
         });
 
-        UpcomingList();
+        TodaysdataList();
+
         return v;
 
     }
@@ -232,7 +237,7 @@ public class UpcGuest extends Fragment {
 
                             String status = j.getString("status");
                             if (status.equals("1")) {
-                                UpcomingList();
+                                TodaysdataList();
                             }
                             else {
                                 Toasty.info(getActivity(), ""+j.getString("messege"),
@@ -270,9 +275,9 @@ public class UpcGuest extends Fragment {
         mRequestQueue1.add(stringRequest1);
     }
 
-    public void UpcomingList() {
+    public void TodaysdataList() {
         mainscreen.setVisibility(View.VISIBLE);
-        String url = Config.upcominglist;
+        String url = Config.todayslist;
         mRequestQueue = Volley.newRequestQueue(getActivity());
 
         stringRequest = new StringRequest(Request.Method.POST, url,
@@ -287,8 +292,7 @@ public class UpcGuest extends Fragment {
 
                             String status = j.getString("status");
                             if (status.equals("1")) {
-
-                                JSONArray applist = j.getJSONArray("Upcoming_guests");
+                                JSONArray applist = j.getJSONArray("todays_guests");
                                 if (applist != null && applist.length() > 0) {
                                     for (int i = 0; i < applist.length(); i++) {
 //                               for (int i = 0; i <= 10; i++) {
@@ -302,18 +306,19 @@ public class UpcGuest extends Fragment {
                                         model.setVstpurpse(getOne.getString("visit_purpose"));
                                         model.setDatess(getOne.getString("visit_date"));
                                         model.setGuestrole(getOne.getString("guest_role"));
+                                        model.setGueststatus(getOne.getString("guest_status"));
                                         model.setSettime(getOne.getString("visit_time"));
+                                        model.setIntime(getOne.getString("checkin_time"));
                                         model.setTtlguest(getOne.getString("total_guest"));
 
                                         list.add(model);
-                                        upcGuestAdapter = new UpcGuestAdapter(getActivity(), list);
-                                        recyclerView.setAdapter(upcGuestAdapter);
+                                        todaysGuestAdapter = new TodaysGuestAdapter(getActivity(), list);
+                                        recyclerView.setAdapter(todaysGuestAdapter);
                                         nodata.setVisibility(View.GONE);
                                         recyclerView.setVisibility(View.VISIBLE);
                                     }
                                 }
-                            }
-                            else {
+                            } else {
                                 nodata.setVisibility(View.VISIBLE);
                                 recyclerView.setVisibility(View.GONE);
                                 mainscreen.setVisibility(View.GONE);
@@ -323,12 +328,13 @@ public class UpcGuest extends Fragment {
                         }
                         mainscreen.setVisibility(View.GONE);
                     }
+
                 },
                 new Response.ErrorListener() {
                     @Override
                     public void onErrorResponse(VolleyError error) {
                         progressBar.setVisibility(View.GONE);
-                      /*  Toast toast = Toast.makeText(getActivity(), "Network Error", Toast.LENGTH_SHORT);
+                       /* Toast toast = Toast.makeText(getActivity(), "Network Error", Toast.LENGTH_SHORT);
                         toast.setGravity(Gravity.CENTER, 0, 0);
                         toast.show();*/
                         mainscreen.setVisibility(View.GONE);
@@ -363,11 +369,12 @@ public class UpcGuest extends Fragment {
             @Override
             public void onClick(View view) {
                 dialogs.dismiss();
-                UpcomingList();
+                TodaysdataList();
             }
         });
         dialogs.show();
     }
+
 
     private void bottomsheetdialog(int posy) {
 
@@ -396,12 +403,15 @@ public class UpcGuest extends Fragment {
         dcntct.setText(model.getMobilenm());
         dguest.setText(model.getTtlguest() + " GUEST");
         dpurpose.setText(model.getVstpurpse());
+
         if (model.getGuestrole().equals("1")) {
             guestroles.setImageDrawable(ContextCompat.getDrawable(getActivity(), R.drawable.ic_thumbup));
         } else {
             guestroles.setImageDrawable(ContextCompat.getDrawable(getActivity(), R.drawable.ic_thumbdown));
         }
-
+        if (model.gueststatus.equals("2")) {
+            expstmp.setVisibility(View.VISIBLE);
+        }
        /* cancel.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -419,7 +429,6 @@ public class UpcGuest extends Fragment {
         bottomSheetDialog.show();
     }
 
-
     public String convrttime(String times) {
         SimpleDateFormat dateFormat = new SimpleDateFormat("hh:mm");
         SimpleDateFormat dateFormat2 = new SimpleDateFormat("hh:mmaa");
@@ -432,4 +441,6 @@ public class UpcGuest extends Fragment {
         }
         return times;
     }
+
+
 }
