@@ -1,17 +1,17 @@
-package com.wk.guestpass.app.Fragments;
+package com.wk.guestpass.app.fragments;
 
 import android.app.Dialog;
 import android.app.ProgressDialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SearchView;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.core.content.ContextCompat;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.SearchView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -31,19 +31,18 @@ import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.wk.guestpass.app.FragAdapters.TodaysGuestAdapter;
-import com.wk.guestpass.app.Models.ListModel;
+import com.wk.guestpass.app.fragAdapters.UpcGuestAdapter;
+import com.wk.guestpass.app.models.ListModel;
 import com.wk.guestpass.app.R;
-import com.wk.guestpass.app.Utilities.Config;
-import com.wk.guestpass.app.Utilities.RecyclerTouchListener;
-import com.wk.guestpass.app.Utilities.SessionManager;
+import com.wk.guestpass.app.utilities.Config;
+import com.wk.guestpass.app.utilities.RecyclerTouchListener;
+import com.wk.guestpass.app.utilities.SessionManager;
 import com.github.ybq.android.spinkit.style.CubeGrid;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import com.crashlytics.android.Crashlytics;
-import io.fabric.sdk.android.Fabric;
+
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -51,19 +50,20 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
+import com.crashlytics.android.Crashlytics;
+import io.fabric.sdk.android.Fabric;
 import es.dmoral.toasty.Toasty;
 import pl.droidsonroids.gif.GifImageView;
 
-public class TodaysGuest extends Fragment {
+public class UpcGuest extends Fragment {
 
-    private RecyclerView recyclerView;
     private SessionManager session;
-    private String usersssid, userna;
-    private TextView dname, ddate, dsetime, dintime, dcntct, dguest, dpurpose;
-    private TodaysGuestAdapter todaysGuestAdapter;
+    private String usersssid;
+    private RecyclerView recyclerView;
+    private UpcGuestAdapter upcGuestAdapter;
     SwipeRefreshLayout mSwipeRefreshLayout;
     private ImageView nodata, logout, back, guestroles, expstmp;
+    private TextView dname, ddate, dsetime, dintime, dcntct, dguest, dpurpose;
     private List<ListModel> list = new ArrayList<>();
     private Dialog bottomSheetDialog;
     public static final String TAG = "MyTag";
@@ -79,19 +79,20 @@ public class TodaysGuest extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         Fabric.with(getActivity(), new Crashlytics());
-        View v = inflater.inflate(R.layout.fragtodaysguest, container, false);
-
+        View v = inflater.inflate(R.layout.fragupcominguest, container, false);
         session = new SessionManager(getActivity());
+
         HashMap<String, String> users = session.getUserDetails();
         usersssid = users.get(SessionManager.KEY_ID);
-
         nodata = (ImageView)v.findViewById(R.id.nodaata);
+        logout = v.findViewById(R.id.logout);
+        back  = v.findViewById(R.id.back);
+
         progressBar = (ProgressBar)v. findViewById(R.id.progress);
         mainscreen=(RelativeLayout)v.findViewById(R.id.overmain);
         bgrnd=(RelativeLayout)v.findViewById(R.id.bgrnd);
 
-        logout = v.findViewById(R.id.logout);
-        back  = v.findViewById(R.id.back);
+
         searchView = (SearchView)v.findViewById(R.id.searchview);
         mSwipeRefreshLayout = (SwipeRefreshLayout) v.findViewById(R.id.swipeToRefresh);
         back.setOnClickListener(new View.OnClickListener() {
@@ -101,9 +102,9 @@ public class TodaysGuest extends Fragment {
             }
         });
 
-        recyclerView = (RecyclerView)v.findViewById(R.id.TodaysRecycler);
+        recyclerView = (RecyclerView)v.findViewById(R.id.UpcomingRecycler);
         recyclerView.setHasFixedSize(false);
-       /* RecyclerView.LayoutManager mLayoutManager = new LinearLayoutManager(getActivity());
+     /*   RecyclerView.LayoutManager mLayoutManager = new LinearLayoutManager(getActivity());
         recyclerView.setLayoutManager(mLayoutManager);*/
         recyclerView.setLayoutManager(new GridLayoutManager(getActivity(), 2));
         recyclerView.setItemAnimator(new DefaultItemAnimator());
@@ -111,18 +112,16 @@ public class TodaysGuest extends Fragment {
                 recyclerView, new RecyclerTouchListener.ClickListener() {
             @Override
             public void onClick(View view, int position) {
-                bottomsheetdialog(position);
+                  bottomsheetdialog(position);
             }
 
             @Override
             public void onLongClick(View view, final int position) {
                 final String stat=list.get(position).getGueststatus();
-                if(stat.equals("0")){
                     android.app.AlertDialog.Builder builder = new android.app.AlertDialog.Builder(getActivity());
                     builder.setMessage("Are you sure you want to delete");
                     builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
                         public void onClick(DialogInterface dialog, int id) {
-
                             deleteguest(list.get(position).getIds());
                             dialog.dismiss();
                         }
@@ -135,21 +134,10 @@ public class TodaysGuest extends Fragment {
                     android.app.AlertDialog alertDialog = builder.create();
                     alertDialog.show();
                 }
-                else{
-                    Toasty.error(getActivity(), "Sorry!!! Data Can't be Deleted", Toast.LENGTH_SHORT, true).show();
-                }
-            }
-        }));
-        todaysGuestAdapter=new TodaysGuestAdapter(getActivity(), list);
-        recyclerView.setAdapter(todaysGuestAdapter);
 
-        if(!usersssid.isEmpty()){
-            cubeGrid = new CubeGrid();
-            cubeGrid.setColor(getResources().getColor(R.color.colorPrimary));
-            cubeGrid.start();
-            progressBar.setIndeterminateDrawable(cubeGrid);
-            TodaysdataList();
-        }
+        }));
+        upcGuestAdapter=new UpcGuestAdapter(getActivity(), list);
+        recyclerView.setAdapter(upcGuestAdapter);
 
         mSwipeRefreshLayout = (SwipeRefreshLayout)v.findViewById(R.id.swipeToRefresh);
         mSwipeRefreshLayout.setColorSchemeResources(R.color.colorAccent);
@@ -157,10 +145,18 @@ public class TodaysGuest extends Fragment {
             @Override
             public void onRefresh() {
 
-                TodaysdataList();
+                UpcomingList();
                 mSwipeRefreshLayout.setRefreshing(false);
             }
         });
+
+        if(!usersssid.isEmpty()){
+            cubeGrid = new CubeGrid();
+            cubeGrid.setColor(getResources().getColor(R.color.colorPrimary));
+            cubeGrid.start();
+            progressBar.setIndeterminateDrawable(cubeGrid);
+            UpcomingList();
+        }
 
 
         logout.setOnClickListener(new View.OnClickListener() {
@@ -200,17 +196,16 @@ public class TodaysGuest extends Fragment {
 
             @Override
             public boolean onQueryTextChange(String newText) {
-                if (todaysGuestAdapter == null) {
+                if (upcGuestAdapter == null) {
                     return false;
                 } else {
-                    todaysGuestAdapter.getFilter().filter(newText);
+                    upcGuestAdapter.getFilter().filter(newText);
                     return true;
                 }
             }
         });
 
-        TodaysdataList();
-
+        UpcomingList();
         return v;
 
     }
@@ -237,7 +232,7 @@ public class TodaysGuest extends Fragment {
 
                             String status = j.getString("status");
                             if (status.equals("1")) {
-                                TodaysdataList();
+                                UpcomingList();
                             }
                             else {
                                 Toasty.info(getActivity(), ""+j.getString("messege"),
@@ -275,9 +270,9 @@ public class TodaysGuest extends Fragment {
         mRequestQueue1.add(stringRequest1);
     }
 
-    public void TodaysdataList() {
+    public void UpcomingList() {
         mainscreen.setVisibility(View.VISIBLE);
-        String url = Config.todayslist;
+        String url = Config.upcominglist;
         mRequestQueue = Volley.newRequestQueue(getActivity());
 
         stringRequest = new StringRequest(Request.Method.POST, url,
@@ -292,7 +287,8 @@ public class TodaysGuest extends Fragment {
 
                             String status = j.getString("status");
                             if (status.equals("1")) {
-                                JSONArray applist = j.getJSONArray("todays_guests");
+
+                                JSONArray applist = j.getJSONArray("Upcoming_guests");
                                 if (applist != null && applist.length() > 0) {
                                     for (int i = 0; i < applist.length(); i++) {
 //                               for (int i = 0; i <= 10; i++) {
@@ -306,19 +302,18 @@ public class TodaysGuest extends Fragment {
                                         model.setVstpurpse(getOne.getString("visit_purpose"));
                                         model.setDatess(getOne.getString("visit_date"));
                                         model.setGuestrole(getOne.getString("guest_role"));
-                                        model.setGueststatus(getOne.getString("guest_status"));
                                         model.setSettime(getOne.getString("visit_time"));
-                                        model.setIntime(getOne.getString("checkin_time"));
                                         model.setTtlguest(getOne.getString("total_guest"));
 
                                         list.add(model);
-                                        todaysGuestAdapter = new TodaysGuestAdapter(getActivity(), list);
-                                        recyclerView.setAdapter(todaysGuestAdapter);
+                                        upcGuestAdapter = new UpcGuestAdapter(getActivity(), list);
+                                        recyclerView.setAdapter(upcGuestAdapter);
                                         nodata.setVisibility(View.GONE);
                                         recyclerView.setVisibility(View.VISIBLE);
                                     }
                                 }
-                            } else {
+                            }
+                            else {
                                 nodata.setVisibility(View.VISIBLE);
                                 recyclerView.setVisibility(View.GONE);
                                 mainscreen.setVisibility(View.GONE);
@@ -328,13 +323,12 @@ public class TodaysGuest extends Fragment {
                         }
                         mainscreen.setVisibility(View.GONE);
                     }
-
                 },
                 new Response.ErrorListener() {
                     @Override
                     public void onErrorResponse(VolleyError error) {
                         progressBar.setVisibility(View.GONE);
-                       /* Toast toast = Toast.makeText(getActivity(), "Network Error", Toast.LENGTH_SHORT);
+                      /*  Toast toast = Toast.makeText(getActivity(), "Network Error", Toast.LENGTH_SHORT);
                         toast.setGravity(Gravity.CENTER, 0, 0);
                         toast.show();*/
                         mainscreen.setVisibility(View.GONE);
@@ -369,12 +363,11 @@ public class TodaysGuest extends Fragment {
             @Override
             public void onClick(View view) {
                 dialogs.dismiss();
-                TodaysdataList();
+                UpcomingList();
             }
         });
         dialogs.show();
     }
-
 
     private void bottomsheetdialog(int posy) {
 
@@ -403,15 +396,12 @@ public class TodaysGuest extends Fragment {
         dcntct.setText(model.getMobilenm());
         dguest.setText(model.getTtlguest() + " GUEST");
         dpurpose.setText(model.getVstpurpse());
-
         if (model.getGuestrole().equals("1")) {
             guestroles.setImageDrawable(ContextCompat.getDrawable(getActivity(), R.drawable.ic_thumbup));
         } else {
             guestroles.setImageDrawable(ContextCompat.getDrawable(getActivity(), R.drawable.ic_thumbdown));
         }
-        if (model.gueststatus.equals("2")) {
-            expstmp.setVisibility(View.VISIBLE);
-        }
+
        /* cancel.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -429,6 +419,7 @@ public class TodaysGuest extends Fragment {
         bottomSheetDialog.show();
     }
 
+
     public String convrttime(String times) {
         SimpleDateFormat dateFormat = new SimpleDateFormat("hh:mm");
         SimpleDateFormat dateFormat2 = new SimpleDateFormat("hh:mmaa");
@@ -441,6 +432,4 @@ public class TodaysGuest extends Fragment {
         }
         return times;
     }
-
-
 }
