@@ -1,7 +1,7 @@
-package com.wk.guestpass.app.Utilities;
+package com.wk.guestpass.app.utilities;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.GestureDetector;
 import android.view.MotionEvent;
 import android.view.View;
